+++ example/lib/flutter_map_mbtiles/page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:flutter_map_pmtiles/flutter_map_pmtiles.dart';
+import 'package:flutter_map_mbtiles/flutter_map_mbtiles.dart';
 import 'package:latlong2/latlong.dart';
 
 class FlutterMapPmTilesPage extends StatefulWidget {
@@ -10,45 +10,53 @@ class FlutterMapPmTilesPage extends StatefulWidget {
   State<FlutterMapPmTilesPage> createState() => _FlutterMapPmTilesPageState();
 }
 
-// TODO: use your own tile source https://docs.protomaps.com/pmtiles/cloud-storage
-const tileSource =
-    'https://raw.githubusercontent.com/protomaps/PMTiles/main/spec/v3/stamen_toner(raster)CC-BY%2BODbL_z3.pmtiles';
-
 class _FlutterMapPmTilesPageState extends State<FlutterMapPmTilesPage> {
-  final Future<PmTilesTileProvider> _futureTileProvider =
-      PmTilesTileProvider.fromSource(tileSource);
+  final mbtiles = MBTiles(
+    mbtilesPath: 'assets/mbtiles/countries-raster.mbtiles',
+  );
 
   @override
   Widget build(BuildContext context) {
+    final metadata = mbtiles.getMetadata();
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.white,
-        title: const Text('flutter_map_pmtiles'),
+        title: const Text('flutter_map_mbtiles'),
       ),
-      body: FutureBuilder<PmTilesTileProvider>(
-        future: _futureTileProvider,
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            final tileProvider = snapshot.data!;
-            return FlutterMap(
+      body: Column(
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.all(12),
+            child: Text(
+              'MBTiles Name: ${metadata.name}, '
+                  'Format: ${metadata.format}',
+            ),
+          ),
+          Expanded(
+            child: FlutterMap(
               options: const MapOptions(
-                initialZoom: 1,
-                initialCenter: LatLng(0, 0),
-                maxZoom: 3.49,
+                minZoom: 0,
+                maxZoom: 6,
+                initialZoom: 2,
+                initialCenter: LatLng(49, 9),
               ),
               children: [
-                TileLayer(tileProvider: tileProvider),
+                TileLayer(
+                  tileProvider: MbTilesTilesProvider(
+                    mbtiles: mbtiles,
+                  ),
+                ),
               ],
-            );
-          }
-          if (snapshot.hasError) {
-            debugPrint(snapshot.error.toString());
-            debugPrintStack(stackTrace: snapshot.stackTrace);
-            return Center(child: Text(snapshot.error.toString()));
-          }
-          return const Center(child: CircularProgressIndicator());
-        },
+            ),
+          ),
+        ],
       ),
     );
+  }
+
+  @override
+  void dispose() {
+    mbtiles.dispose();
+    super.dispose();
   }
 }
