+++ example/lib/vector_map_tiles_mbtiles/page.dart
@@ -1,57 +1,72 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map_mbtiles/flutter_map_mbtiles.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:vector_map_tiles/vector_map_tiles.dart';
 import 'package:vector_map_tiles_mbtiles/vector_map_tiles_pmtiles.dart';
 import 'package:vector_tile_renderer/vector_tile_renderer.dart';
 
-// It could be that the hosted PMTiles file is no longer available.
-// Check https://maps.protomaps.com/builds/ to get an up to date build.
-// TODO: use your own tile source https://docs.protomaps.com/pmtiles/cloud-storage
-const tileSource = 'https://build.protomaps.com/20240128.pmtiles';
+class VectorMapTilesPmTilesPage extends StatefulWidget {
+  const VectorMapTilesPmTilesPage({super.key});
 
-class VectorMapTilesPmTilesPage extends StatelessWidget {
-  final Future<MbTilesVectorTileProvider> _futureTileProvider =
-      MbTilesVectorTileProvider.fromSource(tileSource);
+  @override
+  State<VectorMapTilesPmTilesPage> createState() =>
+      _VectorMapTilesPmTilesPageState();
+}
 
-  VectorMapTilesPmTilesPage({super.key});
+class _VectorMapTilesPmTilesPageState extends State<VectorMapTilesPmTilesPage> {
+  final mbtiles = MBTiles(
+    mbtilesPath: 'assets/mbtiles/countries-vector.mbtiles',
+    isPBF: true, // optional, but small performance benefit
+  );
 
   @override
   Widget build(BuildContext context) {
+    final metadata = mbtiles.getMetadata();
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.white,
-        title: const Text('vector_map_tiles_pmtiles'),
+        title: const Text('vector_map_tiles_mbtiles'),
       ),
-      body: FutureBuilder<MbTilesVectorTileProvider>(
-        future: _futureTileProvider,
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            final tileProvider = snapshot.data!;
-            return FlutterMap(
+      body: Column(
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.all(12),
+            child: Text(
+              'MBTiles Name: ${metadata.name}, '
+                  'Format: ${metadata.format}',
+            ),
+          ),
+          Expanded(
+            child: FlutterMap(
               options: const MapOptions(
-                initialZoom: 1,
-                initialCenter: LatLng(0, 0),
-                maxZoom: 3.49,
+                minZoom: 0,
+                maxZoom: 6,
+                initialZoom: 2,
+                initialCenter: LatLng(49, 9),
               ),
               children: [
                 VectorTileLayer(
                   theme: ProvidedThemes.lightTheme(),
                   tileProviders: TileProviders({
-                    'openmaptiles': tileProvider,
+                    'openmaptiles': MbTilesVectorTileProvider(
+                      maxZoom: 6,
+                      minZoom: 1,
+                      mbtiles: mbtiles,
+                    ),
                   }),
                 ),
               ],
-            );
-          }
-          if (snapshot.hasError) {
-            debugPrint(snapshot.error.toString());
-            debugPrintStack(stackTrace: snapshot.stackTrace);
-            return Center(child: Text(snapshot.error.toString()));
-          }
-          return const Center(child: CircularProgressIndicator());
-        },
+            ),
+          ),
+        ],
       ),
     );
+  }
+
+  @override
+  void dispose() {
+    mbtiles.dispose();
+    super.dispose();
   }
 }
