+++ example/lib/vector_map_tiles_pmtiles/page.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:vector_map_tiles/vector_map_tiles.dart';
-import 'package:vector_map_tiles_mbtiles/vector_map_tiles_pmtiles.dart';
 import 'package:vector_map_tiles_pmtiles/vector_map_tiles_pmtiles.dart';
 import 'package:vector_tile_renderer/vector_tile_renderer.dart';
 
@@ -24,7 +23,7 @@ class VectorMapTilesPmTilesPage extends StatelessWidget {
         backgroundColor: Colors.white,
         title: const Text('vector_map_tiles_pmtiles'),
       ),
-      body: FutureBuilder<MbTilesVectorTileProvider>(
+      body: FutureBuilder<PmTilesVectorTileProvider>(
         future: _futureTileProvider,
         builder: (context, snapshot) {
           if (snapshot.hasData) {
